Decrease flush interval of access logs (#33805)
This is a UX improvement, by making access logs more responsive. There
is a tradeoff in responsiveness in performance; we could go even lower
but I feel 1s strikes the best balance (10s is the default).

The max buffer size is 64k. Each log is ~256 bytes. This gives us 263 logs per flush if the internal doesn't triger it
If we have 1 QPS: we would write 6/minute before and 60/minute after
If we have 10 QPS: we would write 22/minute before and 60/minute after
If we have 100 QPS: we would write 227/minute before and after
If we have 1000 QPS: we would write 2270/minute before and after

So this has no impact on users with >250 QPS.