Use a separate service for the MCS host (#35282)
The initial attempt at adding the Kubernetes Multi-Cluster Services (MCS) ClusterSet host (clusterset.local) revealed that Istio relies heavily on the property that there be 1 hostname per service.

This PR reverts some recent work that added the MCS hostname to `model.Service`. This PR splits out the MCS host into a separate `model.Service` that is synthetically generated by the `ServiceImportCache`. This synthetic service is updated whenever the real service is updated, or whenever there is a change to the associated `ServiceImport` resource.