Enable/Fix multi-arch build for DOCKER_V2_BUILDER (#35965)
When DOCKER_V2_BUILDER=true, the build on arm64 would get an error:
./tools/docker: line 29: out/linux_amd64/docker-builder: cannot execute
binary file: Exec format error

which is due to lack of arch and os information in tools/docker when
building with common/scripts/gobuild.sh(so it would lead to an amd64
binary on arm64 platform).

Now we set the right OS and Arch to enable and fix the building both
on arm64 and amd64.

Signed-off-by: trevor.tao <trevor.tao@arm.com>