validation: add validation for re2 max regex size (#36143)
Currently, regex over 1024 program size will NACK envoy.

This introduces a new validation error if the program *length* exceeds
1024. Length is always less than program size. This ensures that anyone
this validation error fails for is already broken (by NACKs); we must
approximate as golang doesn't expose a program size.

In order to ensure that users with length<1024 but size>1024, we also
bump the max size enforced by envoy. This is set to a number 4x higher
than I was able to make any regex with length<1024, but with some margin
in case I missed some cases where there are larger sizes, while also not
being so high to allow absurdly high regexs in the worst case.

Overall, this should result in:
* All users with working config: continue to work
* Users with NACK and length<1024: will start working
* Users with NACK and length>1024: will start rejecting in webhook