pilot: unify and improve sub/unsub in delta and sotw (#37989)
* pilot: unify and improve sub/unsub in delta and sotw

Currently:
* SotW always does a full push on request
* Delta does a partial push only on subscribe, but not on unsubscribe.
* SotW will never skip a push, and will instead send a push with 0
  resources. Delta will skip a push only for EDS

New behavior:
* Both Delta and SotW will do partial pushes for subscribes.
* Both Delta and SotW, If there are only unsubcribes, we will not send a response at all

This optimizes request cases. For example, say we have 100 listeners and
remove 10. 45s later, those listeners will drain. This results in an XDS
requests with 99 routes, 98, 97 ... 90 routes - about 1000 routes
generated in total. With this optimization, we will send 0.

On the other side, say we add 1 listener. Before, we would push RDS
(100), and then immediately get a RDS request and send all 101 again.
With this change, we would send the same 100 in the initial RDS push,
but on the request we would just send the new 1 requested.

* fix sds

* address comments